[IMP] Add test to document make to order + over delivery problem.
To reproduce the problem in the web interface:

1. Create Odoo demo instance

2. Install sales, purchase and stock.

3. In the general settings switch on Inventory/Products/Units of
   Measure, Inventory/Traceability/Lots & Serial Numbers,
   Inventory/Warehouse/Storage Locations and
   Inventory/Warehouse/Multi-step routes

4. Crete product Salmon. Unit in Kg. Can be sold and purchased. Set
   Buy and Make To Order routes, Tracking by Lots. Set Asustek as
   seller.

5. Creat sale order to deliver 5 Kg. of Salmon to Agrolait.

6. Confirm sale order. Odoo should have created the picking a purchase
   order.

7. Confirm the purchase order and receive the merchandise. Set 000001
   as lot number.

8. Check that the picking of the sale order has 5 Kg. of the 000001
   lot reserved.

9. Transfer the picking but set 6Kg instead of the reserved 5Kg as
   done.

EXPECTED RESULT:

a) The picking is in state done. The done quantity is 5 Kg.

b) A new picking is created in waiting state. It is waiting for 1 Kg.

c) A new purchase order of 1 Kg. is create to satisfy the needs of the
   new picking.

SEEN RESULT:

a) The picking is in state done. The done quantity is 0 Kg. and the
   initial quantity 1 Kg.

   This is incorrect.

b) A new picking is created in the ready state. The reserve is 5Kg and
  the done quantity is 6Kg.

   This is incorrect.

c) A new purchase order of 1 Kg. is create to satisfy the needs of the
   new picking.